Struct aws_sdk_ecr::client::fluent_builders::DescribeImages

pub struct DescribeImages<C = DynConnector> { /* fields omitted */ }

Implementations

impl<C> DescribeImages<C>

pub async fn send(
self
) -> Result<DescribeImagesOutput, SdkError<DescribeImagesError>> where
C: SmithyConnector,

pub fn registry_id(self, input: impl Into<String>) -> Self

The AWS account ID associated with the registry that contains the repository in which to describe images. If you do not specify a registry, the default registry is assumed.

pub fn set_registry_id(self, input: Option<String>) -> Self

pub fn repository_name(self, input: impl Into<String>) -> Self

The repository that contains the images to describe.

pub fn set_repository_name(self, input: Option<String>) -> Self

pub fn image_ids(self, inp: impl Into<ImageIdentifier>) -> Self

The list of image IDs for the requested repository.

pub fn set_image_ids(self, input: Option<Vec<ImageIdentifier>>) -> Self

pub fn next_token(self, input: impl Into<String>) -> Self

The nextToken value returned from a previous paginated DescribeImages request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify images with imageIds.

pub fn set_next_token(self, input: Option<String>) -> Self

pub fn max_results(self, input: i32) -> Self

The maximum number of repository results returned by DescribeImages in paginated output. When this parameter is used, DescribeImages only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImages request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeImages returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify images with imageIds.

pub fn set_max_results(self, input: Option<i32>) -> Self

pub fn filter(self, input: DescribeImagesFilter) -> Self

The filter key and value with which to filter your DescribeImages results.

pub fn set_filter(self, input: Option<DescribeImagesFilter>) -> Self