Module aws_sdk_ecr::error

Modules

empty_upload_exception

See EmptyUploadException

image_already_exists_exception

See ImageAlreadyExistsException

image_digest_does_not_match_exception

See ImageDigestDoesNotMatchException

image_not_found_exception

See ImageNotFoundException

image_tag_already_exists_exception

See ImageTagAlreadyExistsException

invalid_layer_exception

See InvalidLayerException

invalid_layer_part_exception

See InvalidLayerPartException

invalid_parameter_exception

See InvalidParameterException

invalid_tag_parameter_exception

See InvalidTagParameterException

kms_exception

See KmsException

layer_already_exists_exception

See LayerAlreadyExistsException

layer_inaccessible_exception

See LayerInaccessibleException

layer_part_too_small_exception

See LayerPartTooSmallException

layers_not_found_exception

See LayersNotFoundException

lifecycle_policy_not_found_exception

See LifecyclePolicyNotFoundException

lifecycle_policy_preview_in_progress_exception

See LifecyclePolicyPreviewInProgressException

lifecycle_policy_preview_not_found_exception

See LifecyclePolicyPreviewNotFoundException

limit_exceeded_exception

See LimitExceededException

referenced_images_not_found_exception

See ReferencedImagesNotFoundException

registry_policy_not_found_exception

See RegistryPolicyNotFoundException

repository_already_exists_exception

See RepositoryAlreadyExistsException

repository_not_empty_exception

See RepositoryNotEmptyException

repository_not_found_exception

See RepositoryNotFoundException

repository_policy_not_found_exception

See RepositoryPolicyNotFoundException

scan_not_found_exception

See ScanNotFoundException

server_exception

See ServerException

too_many_tags_exception

See TooManyTagsException

unsupported_image_type_exception

See UnsupportedImageTypeException

upload_not_found_exception

See UploadNotFoundException

validation_exception

See ValidationException

Structs

BatchCheckLayerAvailabilityError

BatchDeleteImageError

BatchGetImageError

CompleteLayerUploadError

CreateRepositoryError

DeleteLifecyclePolicyError

DeleteRegistryPolicyError

DeleteRepositoryError

DeleteRepositoryPolicyError

DescribeImageScanFindingsError

DescribeImagesError

DescribeRegistryError

DescribeRepositoriesError

EmptyUploadException

The specified layer upload does not contain any layer parts.

GetAuthorizationTokenError

GetDownloadUrlForLayerError

GetLifecyclePolicyError

GetLifecyclePolicyPreviewError

GetRegistryPolicyError

GetRepositoryPolicyError

ImageAlreadyExistsException

The specified image has already been pushed, and there were no changes to the manifest

ImageDigestDoesNotMatchException

The specified image digest does not match the digest that Amazon ECR calculated for the

ImageNotFoundException

The image requested does not exist in the specified repository.

ImageTagAlreadyExistsException

The specified image is tagged with a tag that already exists. The repository is

InitiateLayerUploadError

InvalidLayerException

The layer digest calculation performed by Amazon ECR upon receipt of the image layer does

InvalidLayerPartException

The layer part size is not valid, or the first byte specified is not consecutive to

InvalidParameterException

The specified parameter is invalid. Review the available parameters for the API

InvalidTagParameterException

An invalid parameter has been specified. Tag keys can have a maximum character length of 128 characters, and tag values can have

KmsException

The operation failed due to a KMS exception.

LayerAlreadyExistsException

The image layer already exists in the associated repository.

LayerInaccessibleException

The specified layer is not available because it is not associated with an image.

LayerPartTooSmallException

Layer parts must be at least 5 MiB in size.

LayersNotFoundException

The specified layers could not be found, or the specified layer is not valid for this

LifecyclePolicyNotFoundException

The lifecycle policy could not be found, and no policy is set to the

LifecyclePolicyPreviewInProgressException

The previous lifecycle policy preview request has not completed. Wait and try

LifecyclePolicyPreviewNotFoundException

There is no dry run for this repository.

LimitExceededException

The operation did not succeed because it would have exceeded a service limit for your

ListImagesError

ListTagsForResourceError

PutImageError

PutImageScanningConfigurationError

PutImageTagMutabilityError

PutLifecyclePolicyError

PutRegistryPolicyError

PutReplicationConfigurationError

ReferencedImagesNotFoundException

The manifest list is referencing an image that does not exist.

RegistryPolicyNotFoundException

The registry doesn't have an associated registry policy.

RepositoryAlreadyExistsException

The specified repository already exists in the specified registry.

RepositoryNotEmptyException

The specified repository contains images. To delete a repository that contains images,

RepositoryNotFoundException

The specified repository could not be found. Check the spelling of the specified

RepositoryPolicyNotFoundException

The specified repository and registry combination does not have an associated

ScanNotFoundException

The specified image scan could not be found. Ensure that image scanning is enabled on

ServerException

These errors are usually caused by a server-side issue.

SetRepositoryPolicyError

StartImageScanError

StartLifecyclePolicyPreviewError

TagResourceError

TooManyTagsException

The list of tags on the repository is over the limit. The maximum number of tags that

UnsupportedImageTypeException

The image is of a type that cannot be scanned.

UntagResourceError

UploadLayerPartError

UploadNotFoundException

The upload could not be found, or the specified upload ID is not valid for this

ValidationException

There was an exception validating this request.

Enums

BatchCheckLayerAvailabilityErrorKind

BatchDeleteImageErrorKind

BatchGetImageErrorKind

CompleteLayerUploadErrorKind

CreateRepositoryErrorKind

DeleteLifecyclePolicyErrorKind

DeleteRegistryPolicyErrorKind

DeleteRepositoryErrorKind

DeleteRepositoryPolicyErrorKind

DescribeImageScanFindingsErrorKind

DescribeImagesErrorKind

DescribeRegistryErrorKind

DescribeRepositoriesErrorKind

GetAuthorizationTokenErrorKind

GetDownloadUrlForLayerErrorKind

GetLifecyclePolicyErrorKind

GetLifecyclePolicyPreviewErrorKind

GetRegistryPolicyErrorKind

GetRepositoryPolicyErrorKind

InitiateLayerUploadErrorKind

ListImagesErrorKind

ListTagsForResourceErrorKind

PutImageErrorKind

PutImageScanningConfigurationErrorKind

PutImageTagMutabilityErrorKind

PutLifecyclePolicyErrorKind

PutRegistryPolicyErrorKind

PutReplicationConfigurationErrorKind

SetRepositoryPolicyErrorKind

StartImageScanErrorKind

StartLifecyclePolicyPreviewErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UploadLayerPartErrorKind