Module aws_sdk_ecr::model

Modules

attribute

See Attribute

authorization_data

See AuthorizationData

describe_images_filter

See DescribeImagesFilter

encryption_configuration

See EncryptionConfiguration

image

See Image

image_detail

See ImageDetail

image_failure

See ImageFailure

image_identifier

See ImageIdentifier

image_scan_finding

See ImageScanFinding

image_scan_findings

See ImageScanFindings

image_scan_findings_summary

See ImageScanFindingsSummary

image_scan_status

See ImageScanStatus

image_scanning_configuration

See ImageScanningConfiguration

layer

See Layer

layer_failure

See LayerFailure

lifecycle_policy_preview_filter

See LifecyclePolicyPreviewFilter

lifecycle_policy_preview_result

See LifecyclePolicyPreviewResult

lifecycle_policy_preview_summary

See LifecyclePolicyPreviewSummary

lifecycle_policy_rule_action

See LifecyclePolicyRuleAction

list_images_filter

See ListImagesFilter

replication_configuration

See ReplicationConfiguration

replication_destination

See ReplicationDestination

replication_rule

See ReplicationRule

repository

See Repository

tag

See Tag

Structs

Attribute

This data type is used in the ImageScanFinding data type.

AuthorizationData

An object representing authorization data for an Amazon ECR registry.

DescribeImagesFilter

An object representing a filter on a DescribeImages

EncryptionConfiguration

The encryption configuration for the repository. This determines how the contents of

Image

An object representing an Amazon ECR image.

ImageDetail

An object that describes an image returned by a DescribeImages

ImageFailure

An object representing an Amazon ECR image failure.

ImageIdentifier

An object with identifying information for an Amazon ECR image.

ImageScanFinding

Contains information about an image scan finding.

ImageScanFindings

The details of an image scan.

ImageScanFindingsSummary

A summary of the last completed image scan.

ImageScanStatus

The current status of an image scan.

ImageScanningConfiguration

The image scanning configuration for a repository.

Layer

An object representing an Amazon ECR image layer.

LayerFailure

An object representing an Amazon ECR image layer failure.

LifecyclePolicyPreviewFilter

The filter for the lifecycle policy preview.

LifecyclePolicyPreviewResult

The result of the lifecycle policy preview.

LifecyclePolicyPreviewSummary

The summary of the lifecycle policy preview request.

LifecyclePolicyRuleAction

The type of action to be taken.

ListImagesFilter

An object representing a filter on a ListImages operation.

ReplicationConfiguration

The replication configuration for a registry.

ReplicationDestination

An array of objects representing the details of a replication destination.

ReplicationRule

An array of objects representing the replication destinations for a replication

Repository

An object representing a repository.

Tag

The metadata that you apply to a resource to help you categorize and organize them.

Enums

EncryptionType

FindingSeverity

ImageActionType

ImageFailureCode

ImageTagMutability

LayerAvailability

LayerFailureCode

LifecyclePolicyPreviewStatus

ScanStatus

TagStatus