Module aws_sdk_ecr::operation

Structs

BatchCheckLayerAvailability

Checks the availability of one or more image layers in a repository.

BatchDeleteImage

Deletes a list of specified images within a repository. Images are specified with

BatchGetImage

Gets detailed information for an image. Images are specified with either an

CompleteLayerUpload

Informs Amazon ECR that the image layer upload has completed for a specified registry,

CreateRepository

Creates a repository. For more information, see Amazon ECR Repositories in the

DeleteLifecyclePolicy

Deletes the lifecycle policy associated with the specified repository.

DeleteRegistryPolicy

Deletes the registry permissions policy.

DeleteRepository

Deletes a repository. If the repository contains images, you must either delete all

DeleteRepositoryPolicy

Deletes the repository policy associated with the specified repository.

DescribeImageScanFindings

Returns the scan findings for the specified image.

DescribeImages

Returns metadata about the images in a repository.

DescribeRegistry

Describes the settings for a registry. The replication configuration for a repository

DescribeRepositories

Describes image repositories in a registry.

GetAuthorizationToken

Retrieves an authorization token. An authorization token represents your IAM

GetDownloadUrlForLayer

Retrieves the pre-signed Amazon S3 download URL corresponding to an image layer. You

GetLifecyclePolicy

Retrieves the lifecycle policy for the specified repository.

GetLifecyclePolicyPreview

Retrieves the results of the lifecycle policy preview request for the specified

GetRegistryPolicy

Retrieves the permissions policy for a registry.

GetRepositoryPolicy

Retrieves the repository policy for the specified repository.

InitiateLayerUpload

Notifies Amazon ECR that you intend to upload an image layer.

ListImages

Lists all the image IDs for the specified repository.

ListTagsForResource

List the tags for an Amazon ECR resource.

PutImage

Creates or updates the image manifest and tags associated with an image.

PutImageScanningConfiguration

Updates the image scanning configuration for the specified repository.

PutImageTagMutability

Updates the image tag mutability settings for the specified repository. For more

PutLifecyclePolicy

Creates or updates the lifecycle policy for the specified repository. For more

PutRegistryPolicy

Creates or updates the permissions policy for your registry.

PutReplicationConfiguration

Creates or updates the replication configuration for a registry. The existing

SetRepositoryPolicy

Applies a repository policy to the specified repository to control access permissions.

StartImageScan

Starts an image vulnerability scan. An image scan can only be started once per day on

StartLifecyclePolicyPreview

Starts a preview of a lifecycle policy for the specified repository. This allows you

TagResource

Adds specified tags to a resource with the specified ARN. Existing tags on a resource

UntagResource

Deletes specified tags from a resource.

UploadLayerPart

Uploads an image layer part to Amazon ECR.